fix 参数解析器注册不上
